+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 2.4.0
 ### Planned
 * [x] GOG Parser.
-* [ ] UPlay Parser.
+* [x] UPlay Parser.
 * [ ] Origin Parser.
 * [ ] Command Line Interface.
 * [x] Capability to override image searches.
@@ -12,6 +12,19 @@ All notable changes to this project will be documented in this file.
 * [x] Automatic Preset Updating.
 * [x] Make icons work for Steam parser (with the caveat that they will most likely be overwritten when a developer updates their game).
 
+## 2.3.33
+### Fixed
+* Actually fixed the title sanitization issue in `screenshots.vdf`. [Issue 296](https://github.com/SteamGridDB/steam-rom-manager/issues/296).
+* Fixed [Issue 284](https://github.com/SteamGridDB/steam-rom-manager/issues/284) and [Issue 278](https://github.com/SteamGridDB/steam-rom-manager/issues/278)
+
+## 2.3.32
+### Added
+* UPlay Parser.
+### Changed
+* Prettied up the settings page further.
+### Fixed
+* Now properly sanitizing the titles put into `screenshots.vdf`.
+* Fixed the change in 2.3.31 that prevented GOG/Epic Parser from working unless directory overrides were set.
 
 ## 2.3.31
 ### Added